fix : error fix in login & hide scroll bar
