+++ src/main/frontend/package.json
@@ -30,15 +30,14 @@
     "next": "^15.0.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-markdown": "^9.0.1",
     "react-router-dom": "^7.0.2",
     "react-scripts": "^5.0.1",
     "react-slick": "^0.30.2",
     "react-spinners": "^0.14.1",
-    "react-markdown": "^9.0.1",
-    "react-scripts": "^5.0.1",
-    "react-slick": "^0.30.2",
     "remark-gfm": "^4.0.0",
     "slick-carousel": "^1.8.1",
+    "tailwind-scrollbar-hide": "^1.1.7",
     "tailwindcss": "^3.4.13",
     "typescript": "^5.6.2",
     "web-vitals": "^2.1.4",
