+++ src/main/frontend/tailwind.config.ts
@@ -90,7 +90,8 @@ const config: Config = {
   plugins: [
     require('@tailwindcss/forms'),
     require('@tailwindcss/typography'),
-    require('@tailwindcss/line-clamp')
+    require('@tailwindcss/line-clamp'),
+    require('tailwind-scrollbar-hide'),
   ],
 };
 export default config;
